requirements.txt
confluent-kafka==2.3.0
mysql-connector-python==8.0.33